requirements.txt
flask>=2.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
selenium>=4.0.0
werkzeug>=2.0.0
watchdog>=2.1.0
openai>=1.0.0
python-dotenv>=1.0.0 